remove warnings added dummy json files for bluesim and vc707 targets updated dummy files to have 'fmc' content added dummy examples/hdmidisplay/hdmidisplay-vc707.xdc file

+++ examples/echo/testecho.cpp
@@ -97,7 +97,7 @@ static void call_say(int v)
     PREPAREWAIT(sem_heard2);
     echoRequestProxy->say(v);
     SEMWAIT(&sem_heard2);
-    printf("call_say: elapsed %" PRIu64 "\n", LAP_TIMER(0));
+    printf("call_say: elapsed %llu\n", (unsigned long long)LAP_TIMER(0));
 }
 
 static void call_say2(int v, int v2)
